Prepare CloudFormation template for import
The original plan was to set up the infrastructure using the console. The UI is pretty helpful in learning AWS. A more long-term solution is to use something like CloudFormation or Terraform. I was going to do a resource import later. It turns out that not all resource types support import and drift detection. So I'll need to take the tricker approach. I'll have to drop everything and recreate it with CloudFormation.

https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/resource-import-supported-resources.html