Get the CloudFormation template to successfully deploy
I deleted the resources that I created from the console and used the template.

It seems like there's no way to validate a CloudFormation template other than to deploy it. The only way to check if a service will accept a request is to send that request to the service to create the thing.

Also, just because it was successfully deployed doesn't mean that it's configured correctly and working.