Allow changes to be pulled while measurements are in edit mode
Staging more often makes this a little bit easier and safer. It means that there's a very narrow time window between the user making a change and it existing in the staged database. If we end up overwriting the user's changes in this situation then so be it. It means they'll have to redo something they did a second ago because they were also making changes on another device for some reason.