Start bringing in Ionic
I was getting frustrated with my own feeble attempts at replicating the iOS look-and-feel. My intention with Tailwind was to come up with my own style so that it didn't seem like I was trying and failing to copy iOS. My plans were foiled by the fact that I'm a programmer, not a designer. I also really like the way that iOS looks so my own style ended up being iOS.

It was getting to a point where I realised that I was falling for the sunk-cost fallacy. I should just bite the bullet and use Ionic. The app will look so much more polished that way.

There's going to be a lot of work ahead to polish this off though. I'm going to phase-out Tailwind, heroicons and vue-final-modal. I'll use Ionic components in more places. Quite a few things are broken at this point.